add history

+++ Cinema.Movie/Cinema.Movie.Web/Modules/Movie/Movie/MovieColumns.cs
@@ -43,7 +43,7 @@ namespace Cinema.Movie.Movie.Columns
         public Int32 Budget { get; set; }
         public List<Int16> GenreList { get; set; }
         public List<string> GenreListName { get; set; }
-        public List<Int16> TagList { get; set; }
+        public List<Int64> TagList { get; set; }
         public List<string> TagListName { get; set; }
         public List<Int32> CountryList { get; set; }
         public List<string> CountryListName { get; set; }
